start backend in seperate worker thread
